moved logic of query setting from search bar to products page component
