+++ src/components/SearchBar/SearchBar.tsx
@@ -1,42 +1,19 @@
 import classNames from 'classnames';
-import React, { useCallback } from 'react';
-import { useSearchParams } from 'react-router-dom';
-import { debounce } from '../../utils/debounce';
+import React from 'react';
 import Icon from '../Icon/Icon';
 import styles from './SearchBar.module.scss';
 
 type Props = {
   query: string;
-  setQuery: (value: string) => void;
-  setIsQueryNew: (value: boolean) => void;
+  changeQuery: (value: string) => void;
 };
 
-export const SearchBar: React.FC<Props> = ({
-  query,
-  setQuery,
-  setIsQueryNew,
-}) => {
-  const [searchParams, setSearchParams] = useSearchParams();
-  const setQueryToApply = (value: string) => {
-    searchParams.set('page', String(1));
-
-    if (!value) {
-      searchParams.delete('search');
-    } else {
-      searchParams.set('search', String(value));
-    }
-
-    setIsQueryNew(true);
-    setSearchParams(searchParams);
-  };
-  const applyQuery = useCallback(debounce(setQueryToApply, 1000), []);
-  const onQueryInput = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setQuery(event.target.value);
-    applyQuery(event.target.value);
-  };
-
+export const SearchBar: React.FC<Props> = ({ query, changeQuery }) => {
   const onFormSubmit = (event: React.FormEvent<HTMLFormElement>) =>
     event.preventDefault();
+
+  const onQueryInput = (event: React.ChangeEvent<HTMLInputElement>) =>
+    changeQuery(event.target.value);
 
   return (
     <form className={styles.search} onSubmit={onFormSubmit}>
